requirements.txt
aniso8601==4.0.1
asn1crypto==0.24.0
astroid==2.1.0
autopep8==1.4.3
cffi==1.11.5
Click==7.0
colorama==0.4.1
cryptography==2.4.2
Flask==1.0.2
Flask-Cors==3.0.7
Flask-RESTful==0.3.6
idna==2.8
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
lxml==4.2.5
MarkupSafe==1.1.0
mccabe==0.6.1
pep8==1.7.1
Pillow==5.3.0
pycodestyle==2.4.0
pycparser==2.19
pylint==2.2.2
PyMySQL==0.9.2
python-docx==0.8.7
pytz==2018.7
six==1.12.0
SQLAlchemy==1.2.15
typed-ast==1.1.0
Werkzeug==0.14.1
wrapt==1.10.11
